requirements.txt
python-telegram-bot>=20.0
aiohttp>=3.9.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
pytz>=2024.1
